+++ node/middlewares/generateMiddlewares/generateProductRoutes.ts
@@ -1,6 +1,6 @@
 import { Binding } from '@vtex/api'
 import { Product } from 'vtex.catalog-graphql'
-import { getBucket, hashString, TENANT_CACHE_TTL_S } from '../../utils'
+import { CONFIG_BUCKET, GENERATION_CONFIG_FILE, getBucket, hashString, TENANT_CACHE_TTL_S } from '../../utils'
 import {
   createFileName,
   createTranslator,
@@ -43,11 +43,13 @@ export async function generateProductRoutes(ctx: EventContext, next: () => Promi
   })
 
   const {
-    authToken,
     from,
+    generationId,
     processedProducts,
     invalidProducts,
   }: ProductRoutesGenerationEvent = body!
+
+  const { authToken } = await vbase.getJSON<GenerationConfig>(CONFIG_BUCKET, GENERATION_CONFIG_FILE)
 
   const to = from + PAGE_LIMIT - 1
   const { data, range: { total } } = await catalog.getProductsAndSkuIds(from, to, authToken)
@@ -120,8 +122,8 @@ export async function generateProductRoutes(ctx: EventContext, next: () => Promi
   )
 
   const payload: ProductRoutesGenerationEvent = {
-    authToken,
     from: from + PAGE_LIMIT,
+    generationId,
     invalidProducts: invalidProducts + currentInvalidProducts,
     processedProducts: processedProducts + currentProcessedProducts,
   }
@@ -140,6 +142,7 @@ export async function generateProductRoutes(ctx: EventContext, next: () => Promi
     ctx.state.nextEvent = {
       event: GROUP_ENTRIES_EVENT,
       payload: {
+        generationId,
         indexFile: PRODUCT_ROUTES_INDEX,
       },
     }