+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fallbacks to first store binding in case of missing binding
+- Increases clients timeouts
+
+### Added
+- Token security, by saving it in VBase
+- Generation Id, to identify a generation and not process events from other ones
 
 ## [2.3.1] - 2020-06-01